+++ source/compute_activity.hpp
@@ -33,7 +33,7 @@ namespace tarsier {
             auto& potential_and_t = _potentials_and_ts[event.x + event.y * _width];
             potential_and_t.first =
                 potential_and_t.first
-                    * std::exp(-static_cast<double>(event.t - potential_and_t.second) / static_cast<double>(_decay))
+                    * std::exp(-static_cast<float>(event.t - potential_and_t.second) / static_cast<float>(_decay))
                 + 1;
             potential_and_t.second = event.t;
             _handle_activity(_event_to_activity(event, potential_and_t.first));
@@ -44,14 +44,14 @@ namespace tarsier {
         const uint64_t _decay;
         EventToActivity _event_to_activity;
         HandleActivity _handle_activity;
-        std::vector<std::pair<double, uint64_t>> _potentials_and_ts;
+        std::vector<std::pair<float, uint64_t>> _potentials_and_ts;
     };
 
     template <typename Event, typename Activity, typename EventToActivity, typename HandleActivity>
     compute_activity<Event, Activity, EventToActivity, HandleActivity> make_compute_activity(
         uint16_t width,
         uint16_t height,
-        double decay,
+        float decay,
         EventToActivity event_to_activity,
         HandleActivity handle_activity) {
         return compute_activity<Event, Activity, EventToActivity, HandleActivity>(
